requirements.txt
requests
python-dotenv
numpy
pillow
pandas
scikit-learn
tensorflow
datasets
transformers
sentencepiece
torch
tf-keras
transformers[torch]